Refactoring

+++ src/main/java/com/select/choice/domain/post/domain/service/impl/PostServiceImpl.java
@@ -7,13 +7,12 @@ import com.select.choice.domain.post.domain.data.dto.CreatePostDto;
 import com.select.choice.domain.post.domain.data.dto.PostDetailDto;
 import com.select.choice.domain.post.domain.data.dto.PostDto;
 import com.select.choice.domain.post.domain.data.entity.Post;
+import com.select.choice.domain.post.domain.exception.IsNotMyPostException;
 import com.select.choice.domain.post.domain.data.response.PostDetailResponse;
 import com.select.choice.domain.post.domain.exception.PostNotFoundException;
 import com.select.choice.domain.post.domain.repository.PostRepository;
 import com.select.choice.domain.post.domain.service.PostService;
 import com.select.choice.domain.post.domain.util.PostConverter;
-import com.select.choice.domain.user.domain.data.entity.User;
-import com.select.choice.domain.user.domain.facade.UserFacade;
 import com.select.choice.domain.user.domain.data.entity.User;
 import com.select.choice.domain.user.domain.facade.UserFacade;
 import com.select.choice.global.error.type.ErrorCode;
@@ -22,6 +21,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -62,7 +62,10 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public void deletePost(Long postIdx) {
+        User user = userFacade.currentUser();
         Post post = postRepository.findById(postIdx).orElseThrow(()->new PostNotFoundException(ErrorCode.POST_NOT_FOUND));
+        if(!Objects.equals(post.getUser().getIdx(), user.getIdx()))
+            throw new IsNotMyPostException(ErrorCode.IS_NOT_MY_POST);
         postRepository.delete(post);
     }
 
@@ -71,7 +74,7 @@ public class PostServiceImpl implements PostService {
     public void addCount(AddCountDto addCountDto, Long postIdx) {
         Post post = postRepository.findById(postIdx).orElseThrow(()->new PostNotFoundException(ErrorCode.POST_NOT_FOUND));
         Integer choice = addCountDto.getChoice();
-        if(choice == 1){
+        if(choice == 0){
             post.updateFirstVotingCount();
         } else
             post.updateSecondVotingCount();
